e2e tests moved to an independent file
